CNS-952: fix lint issue

+++ x/subscription/keeper/cluster.go
@@ -10,7 +10,7 @@ func (k Keeper) GetAllClusters(ctx sdk.Context) []string {
 	plans := k.plansKeeper.GetAllPlanIndices(ctx)
 	for _, plan := range plans {
 		for _, usage := range types.SUB_USAGES {
-			sub := types.Subscription{PlanIndex: plan, DurationTotal: uint64(usage)}
+			sub := types.Subscription{PlanIndex: plan, DurationTotal: usage}
 			cluster := types.GetClusterKey(sub)
 			clusters = append(clusters, cluster)
 		}
